(not sure why there were conflicts, took the versions from the release)

Conflicts:
	engine-spark/src/main/scala/com/intel/intelanalytics/engine/spark/frame/parquet/ParquetReader.scala
	engine-spark/src/main/scala/com/intel/intelanalytics/engine/spark/frame/plugins/AddColumnsPlugin.scala

+++ interfaces/src/main/scala/com/intel/intelanalytics/domain/query/RowQuery.scala
@@ -1,7 +1,7 @@
 //////////////////////////////////////////////////////////////////////////////
 // INTEL CONFIDENTIAL
 //
-// Copyright 2014 Intel Corporation All Rights Reserved.
+// Copyright 2015 Intel Corporation All Rights Reserved.
 //
 // The source code contained or described herein and all documents related to
 // the source code (Material) are owned by Intel Corporation or its suppliers
@@ -31,6 +31,6 @@ package com.intel.intelanalytics.domain.query
  * @param count Number of records to return
  * @tparam Identifier DataType of Primary Key. Will usually be a Long
  */
-case class RowQuery[Identifier](id: Identifier, offset: Long, count: Int) {
+case class RowQuery[Identifier](id: Identifier, offset: Long, count: Long) {
   require(count >= 0, s"Bad count value $count. Count must be >= 0.")
 }
